+++ internal/labels/labels.go
@@ -4,7 +4,6 @@ const (
 	LEnabled    = "dev.polarfox.marina.enabled"
 	LInstanceID = "dev.polarfox.marina.instanceID"
 	LExclude    = "dev.polarfox.marina.exclude"
-	LTags       = "dev.polarfox.marina.tags"
 
 	// Volume
 	LPaths        = "dev.polarfox.marina.paths"